♻️ refactor: inline backendv2 c header helpers
